+++ Structures/ray.hpp
@@ -1,8 +1,5 @@
 #pragma once
 
-#include <vector>
-#include <memory>
-
 #include "point.hpp"
 
 class object;
@@ -16,4 +13,3 @@ struct ray {
 };
 
 auto rayDir(pos_type fov, pos_type x, pos_type y) -> ray;
-auto rayCast(ray r, const std::vector<std::shared_ptr<object>> scene, int bounces) -> point;
