+++ makefile
@@ -5,17 +5,20 @@ CXXFLAGS = -Wall -std=c++20 -g -O3 $(LINK_FLAGS)
 
 all: rt
 
-rt: main.cpp common.hpp objects.o ray.o point.o
-	$(CXX) $(CXXFLAGS) -o rt main.cpp objects.o point.o ray.o
+rt: main.cpp common.hpp objects.o ray.o point.o trace.o
+	$(CXX) $(CXXFLAGS) -o rt main.cpp objects.o point.o ray.o trace.o
 
-objects.o: objects.hpp common.hpp objects.cpp
-	$(CXX) $(CXXFLAGS) -c -o objects.o objects.cpp
+objects.o: Structures/objects.hpp common.hpp Structures/objects.cpp
+	$(CXX) $(CXXFLAGS) -c -o objects.o Structures/objects.cpp
 
-ray.o: ray.hpp ray.cpp common.hpp
-	$(CXX) $(CXXFLAGS) -c -o ray.o ray.cpp
+trace.o: trace.hpp trace.cpp Structures/ray.hpp Structures/objects.hpp common.hpp
+	$(CXX) $(CXXFLAGS) -c -o trace.o trace.cpp
 
-point.o: point.hpp common.hpp point.cpp
-	$(CXX) $(CXXFLAGS) -c -o point.o point.cpp
+ray.o: Structures/ray.hpp Structures/ray.cpp common.hpp
+	$(CXX) $(CXXFLAGS) -c -o ray.o Structures/ray.cpp
+
+point.o: Structures/point.hpp common.hpp Structures/point.cpp
+	$(CXX) $(CXXFLAGS) -c -o point.o Structures/point.cpp
 
 clean:
 	rm *.o
