+++ trace.cpp
@@ -1,27 +1,8 @@
-#include <iostream>
-#include <cstdlib>
-#include <cmath>
+#include "trace.hpp"
 
-#include "ray.hpp"
-#include "objects.hpp"
-#include "point.hpp"
 #include "common.hpp"
-
-[[nodiscard]]
-point ray::p(pos_type t) {
-  // [[assume(t>0)]]
-  return this->e + (this->d * t);
-}
-
-auto rayDir(pos_type fov, pos_type x, pos_type y) -> ray {
-  pos_type halfFov = tan(toRad(90.0 - fov*0.5));
-  pos_type ypart = HEIGHT * 0.5 * halfFov;
-
-  return ray(
-    point(0,0,0),
-    point(x-(WIDTH/2.0), ypart, -(y-HEIGHT/2.0))
-  );
-}
+#include "Structures/objects.hpp"
+#include "Structures/ray.hpp"
 
 auto lightRay(point startpos, const std::vector<std::shared_ptr<object>> scene, int bounces) -> point {
   point colour = point(0,0,0);
